Plot ROC and Related Curves for Phenotyping Models

Plot ROC-like curves to illustrate phenotyping accuracy.

phecap_plot_roc_curves(
  x, axis_x = "1 - spec", axis_y = "sen",
  what = c("training", "random-splits", "validation"),
  ggplot = TRUE, ...)

Arguments

x
either a single object of class PhecapModel or PhecapValidation (returned from phecap_train_phenotyping_model or phecap_validate_phenotyping_model), or a named list of such objects

axis_x
an expression that leads to the x coordinate. Recognized quantities include: cut (probability cutoff), pct (percent of predicted cases), acc (accuracy), tpr (true positive rate), fpr (false positive rate), tnr (true negative rate), ppv (positive predictive value), fdr (false discovery rate), npv (negative predictive value), sen (sensitivity), spec (specificity), prec (precision), rec (recall), f1 (F1 score).

axis_y
an expression that leads to the y coordinate. Recognized quantities are the same as those in axis_x.

what
The curves to be included in the figure.

ggplot
if TRUE and ggplot2 is installed, ggplot will be used for the figure. Otherwise, the base R graphics functions will be used.

…
arguments to be ignored.

See also